show questions in docs/handbook
