+++ gatsby/createPages.js
@@ -232,6 +232,7 @@ module.exports = exports.createPages = async ({ actions, graphql }) => {
                 breadcrumb,
                 breadcrumbBase: { name: 'Handbook', url: '/handbook' },
                 tableOfContents,
+                slug,
             },
         })
     })
@@ -262,6 +263,7 @@ module.exports = exports.createPages = async ({ actions, graphql }) => {
                 breadcrumb,
                 breadcrumbBase: { name: 'Docs', url: '/docs' },
                 tableOfContents,
+                slug,
             },
         })
     })
