+++ src/templates/Handbook/index.js
@@ -12,7 +12,7 @@ import Navigation from './Navigation'
 import '../../styles/handbook.scss'
 
 export default function Handbook({
-    data: { post },
+    data: { post, questions },
     pageContext: { menu, next, previous, breadcrumb = [], breadcrumbBase, tableOfContents },
 }) {
     const { hash } = useLocation()
@@ -95,6 +95,7 @@ export default function Handbook({
                                     next,
                                     previous,
                                     hideLastUpdated,
+                                    questions: questions?.nodes,
                                 }}
                             />
                         </div>
@@ -107,7 +108,7 @@ export default function Handbook({
 }
 
 export const query = graphql`
-    query HandbookQuery($id: String!) {
+    query HandbookQuery($id: String!, $slug: String!) {
         post: mdx(id: { eq: $id }) {
             id
             body
@@ -140,6 +141,16 @@ export const query = graphql`
                         gitLogLatestDate(formatString: "MMM D, YYYY")
                     }
                 }
+            }
+        }
+        questions: allQuestion(filter: { question_slug: { eq: $slug } }) {
+            nodes {
+                question_author
+                question_avatar
+                question_slug
+                question_body
+                answer_author
+                answer_body
             }
         }
     }
